Corrected atomic function in code

+++ arch/powerpc/atomic.h
@@ -317,8 +317,8 @@ atomic_and_and_fetch32 (volatile uint32_t *ptr, uint32_t val)
 }
 
 
-  inline uint32_t 
-atomic_bool_compare_and_swap (volatile uint32_t *ptr, uint32_t oval, uint32_t nval)
+  inline uintptr_t 
+atomic_bool_compare_and_swap (volatile uintptr_t *ptr, uintptr_t oval, uintptr_t nval)
 {
   uint32_t tmp;
